\name{yags.glmReport}
\alias{yags.glmReport}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ report on the GLM component of a GEE solution ... not to
be used directly }
\description{
support for print.yagsResult
}
\usage{
yags.glmReport(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class yagsResult }
  \item{\dots}{ not currently used }
}
\details{
}
\value{
a matrix of model parameter estimates and standard errors
}
\references{ }
\author{ VJ Carey, stvjc@gauss.med.harvard.edu }
\note{ }


\examples{
}
\keyword{ models }
%\keyword{ longitudinal data }% at least one, from doc/KEYWORDS


